+++ sdk/identity/identity/src/msal/utils.ts
@@ -85,12 +85,14 @@ export function getKnownAuthorities(tenantId: string, authorityHost: string): st
  */
 export const defaultLoggerCallback: (
   logger: CredentialLogger,
+  allowPiiLogging?: boolean,
   platform?: "Node" | "Browser"
 ) => msalCommon.ILoggerCallback = (
   logger: CredentialLogger,
+  allowPiiLogging: boolean = false,
   platform: "Node" | "Browser" = isNode ? "Node" : "Browser"
 ) => (level, message, containsPii): void => {
-  if (containsPii) {
+  if (containsPii && !allowPiiLogging) {
     return;
   }
   switch (level) {